requirements.txt
astroid==2.1.0
Django==2.1.5
django-froala-editor==2.9.1.post2
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.4.1
pylint==2.2.2
pytz==2018.9
six==1.12.0
typed-ast==1.2.0
wrapt==1.11.1
